HUGE UPDATE
New bins for janitors to empty. You can hide in them and they have a cool light and a unique sound effect.
The recycler will let you recycle trash for small amounts of sheets of metal and glass. It can be emagged to override the safety, allowing you to gib mobs.

Restructured crates/closets a lot, should be better, maybe a bit confusing.

Misc stuff like papers, the changelog, and organizational things.